+++ netbox_fms/navigation.py
@@ -64,6 +64,25 @@ menu = PluginMenu(
                         ),
                     ),
                 ),
+                PluginMenuItem(
+                    link="plugins:netbox_fms:fiberattenuationspec_list",
+                    link_text="Attenuation Specs",
+                    permissions=["netbox_fms.view_fiberattenuationspec"],
+                    buttons=(
+                        PluginMenuButton(
+                            link="plugins:netbox_fms:fiberattenuationspec_add",
+                            title="Add",
+                            icon_class="mdi mdi-plus-thick",
+                            permissions=["netbox_fms.add_fiberattenuationspec"],
+                        ),
+                        PluginMenuButton(
+                            link="plugins:netbox_fms:fiberattenuationspec_import",
+                            title="Import",
+                            icon_class="mdi mdi-upload",
+                            permissions=["netbox_fms.add_fiberattenuationspec"],
+                        ),
+                    ),
+                ),
             ),
         ),
         (